Tried to Correct Stack Overflow

+++ HumaneSociety/Bird.cs
@@ -8,7 +8,7 @@ namespace HumaneSociety
 {
     public class Bird : Animal
     {
-        public Bird(string animalType, string animalName, float animalWeight, bool hasShots, bool isSick, bool isFed, int kennelNumber, int animalId) : base(animalType, animalName, animalWeight, hasShots, isSick, isFed, kennelNumber, animalId)
+        public Bird(string animalType, string animalName, double animalWeight, bool hasShots, bool isSick, bool isFed, int kennelNumber, int animalId) : base(animalType, animalName, animalWeight, hasShots, isSick, isFed, kennelNumber, animalId)
         {
             animalType = "Bird";
         }
