+++ HumaneSociety/Cat.cs
@@ -8,7 +8,7 @@ namespace HumaneSociety
 {
    public class Cat : Animal
     {
-        public Cat(string animalType, string animalName, float animalWeight, bool hasShots, bool isSick, bool isFed, int kennelNumber, int animalId) : base(animalType, animalName, animalWeight, hasShots, isSick, isFed, kennelNumber, animalId)
+        public Cat(string animalType, string animalName, double animalWeight, bool hasShots, bool isSick, bool isFed, int kennelNumber, int animalId) : base(animalType, animalName, animalWeight, hasShots, isSick, isFed, kennelNumber, animalId)
         {
             animalType = "Cat";
         }
