+++ HumaneSociety/Dog.cs
@@ -9,7 +9,7 @@ namespace HumaneSociety
     public class Dog : Animal
     {
         
-        public Dog (string animalType,string animalName,float animalWeight,bool hasShots,bool isSick, bool isFed, int kennelNumber, int animalId ):base(animalType,animalName,animalWeight,hasShots,isSick,isFed,kennelNumber,animalId)
+        public Dog (string animalType,string animalName,double animalWeight,bool hasShots,bool isSick, bool isFed, int kennelNumber, int animalId ):base(animalType,animalName,animalWeight,hasShots,isSick,isFed,kennelNumber,animalId)
         {
             animalType = "Dog";
         }
